+++ cookbooks/maxlab_plex/metadata.rb
@@ -3,7 +3,7 @@ maintainer 'Maxwell Spangler'
 maintainer_email 'maxcode@maxwellspangler.com'
 license 'All Rights Reserved'
 description 'Installs/Configures maxlab_plex'
-version '1.0.2'
+version '1.1.0'
 chef_version '>= 14.0'
 
 # The `issues_url` points to the location where issues for this cookbook are
@@ -22,4 +22,3 @@ supports 'redhat', '>= 7.0'
 supports 'centos', '>= 7.0'
 
 depends 'maxlab_firewall'
-depends 'selinux_policy'