+++ cookbooks/maxlab_plex/CHANGELOG.md
@@ -2,6 +2,16 @@
 
 Maintain a Plex media server in the maxlab environment.
 
+# 1.1.0
+
+* Enable Test Kitchen development
+* Enable Test Kitchen tests
+* Enable policyfiles
+* When plex is not installed, deploy the plex Linux repo and then install it.
+* When plex is already installed, upgrade it.
+* Use node attributes not a data bag to deploy plex.
+* Marked data bag config_plex with obsolete_
+
 # 1.0.2
 
 * Change tag from 'plex-media-server' to 'plex' for nodes with it deployed.